Remove Laravel
These files have been moved to a branch, at least for now. I'd like
to get back to basics, which is first and foremost, the engine. And
if I try to wrap that in a web app right off the bat, I'm going to
waste time on things that *aren't* the engine.

Note: these files have been moved to a `laravel` branch. If the time
comes, I'll move them back to `main` or I'll just delete the branch.