Commit rudimentary engine
This is version 0.0.1 of the engine. It has two main public methods:

* run() runs the core plan-generation engine
* render() creates output

There are three main internal structures:

* plan (an array of periods with monthly income and expenses
* income (an array of objects describing source income)
* expenses (an array of objects describing monthly expenditures)

In this commit, income with basic monthly, compounding interest is
supported. An opening balance can be specified, with a planned
monthly withdrawal amount, along with APR, and a time for the
income to be tapped (either a fixed year/month or after a dependent
income source is depleted).

Expenses at this point are not implemented. A constant value of 100
is returned for each round of calculations.

This is a good, even if simple, start. The basic ideas are all there
and it will serve as a good place to grow from.