Add incomplete (and unused) Database layer
This can connect to a database and select data from a table, but
that's it. It's also not used at this point, since the information
ultimately intended for the database is hardcoded into PHP data
structures directly in the engine. This will be temporary, while
the basic engine is being built.