Add support for income scenarios
This tool loses a great deal of its usefulness without the ability
to compare different "what if...?" scenarios. With this commit, the
Income class gets scenarios. The previous single-scenario array is
now subdivided into an array of structures, which includes what used
to be the complete "scenario." Further a scenario can specify a
parent scenario, and like-named attributes in a child will be merged
with the parent to allow tweaking rather than requiring each "what
if" to be a fully-formed scenario.