Introduce a MySQL database
At this point the application is still running off the hardcoded
PHP structures for its assets and expenses. But with this commit,
I've taken the first steps toward making this fully database-driven.

I'm using dbdiagram.io for my database design and have added a copy
of its DBML file to this commit. The file retreat-schema.sql is the
export from dbdiagram.io but with minor comments added. There's a
hydration script to fill the empty tables. Then dev-queries.sql
which is a temporary file as I pull all this together.

I already have a few new schema changes lined up. So this is about
to change shortly. But I wanted to capture this very first pass
before continuing.