Update database schema
Main change here is the removal of anything that's a run-time data
construct and not meant to be persisted. So the status tables and
the asset.current_balance column are now gone.

Helper SQL files updated accordingly.

dev-queries.sql also now shows the change from using PHP to merge
parent/child scenarios to using SQL SELECTs to accomplish the same.