Add support for "scenario start period adjustment"
To be honest, I don't know if this one will live. Now that I've got it
working, it just feels "off" to me though I can't put my finger on it
yet. I think in time it will become clear.

Anyway, the idea is that the start period (begin_year + begin_month) is
fixed in the persisted data. But at times it might be nice to override
that at run-time rather than updating a scenario, extending it, or copy/
pasting it to a new scenario. In short, "All things being equal, what
happens if I move the start period up a year." And this will do that,
except for any expenses or assets flagged as "fixed_period" (meaning,
no matter when the simulation begins, these can't be changed: e.g.,
whether I retire in 2024 or 2025 doesn't change the date I turn 59.5)

So, this seems to work (though it lacks heavy testing) but I'm not yet
convinced it's the correct approach or if it will ulimately be more
confusing than useful.