Change cli args so start year/month are optional AND default to db vals
I kept running into this thing where run() takes a user-supplied start
year and start month, but both of these values are also baked into the
scenarios in the database. Why have two sources of starts? I'm leaving
the ability to start on the CLI but with this commit, it will look into
the expense table for the earliest year/month and use that. I may remove
it from the CLI altogether, but I'm not there yet. Still have that issue
with -a --adjustment