Commit Object Refactoring Part 2: Money class
I kept running into issues treating floats as currency, so I created
this simple "money" class to use floats internally, but also ensure
that the values are rounded to two decimal places.

This really isn't in its final form. I'm still a bit conflicted on the
implementation of the comparison operator functions (or if they're
needed at all). But the basics are there, and we're doing well now
and completely avoiding current balances of 0.000000000031177, even
for simple subtraction of 1.00 - 1.00.