Commit Object Refactoring Part 5: The Engine
Lastly comes the Engine refactoring, plus auxiliary changes. Inside the
engine a lot of stuff got cleaned up just due to the breakup of the
monolith. Lost some members, repurposed others, the code got tighter
except where the code got completely removed to other classes.

It's unreasonable to do a blow-by-blow here. Suffice it to say that this
is the line where this little tool crossed from a little experiment and
into real software.  :)

As far as functionality: the core functions from before are all there
and seem to be working fine. One known omission is the full auditing,
which didn't make it completely through the monolith breakup. That will
come next, along with more fixes, tweaks, new features, and improvements.

For now, though, this is a good line in the sand. I'm really happy with
this major refactoring and the promise of where this will go.