Introduce support for a new scenario type: income
Income serves the same purpose as an asset, in that it helps cover
monthly expenses. However, income differs from assets in that it doesn't
draw down from a fixed investment balance. For example, if I derive
income from a job that pays $1000/month for a year, that doesn't mean I
have an asset account with $12,000 in it. It's just money. At first, I
used assets for this purpose, but it was wrong and I had to bite the
bullet and fix it.

In this commit:

* sql file to gather assets
* Income and IncomeCollection classes, mimicking both Assets and Expenses
* Initial unit tests
* Engine update

As far as the Engine goes, after monthly expenses are tallied, it first
attempts to cover expenses from income source(s). If that's enough, it
stops. If more money is still needed, it only then pulls from assets.