Rewrite timeToEnd() logic from scratch
Was having trouble with this in expenses and wasn't fully appreciating
the relationship between the various end periods and infinite entities.
Stepped back, thought about it, and came up with this. After Expenses
worked, applied to Income. Note that income is still nearly a clone of
Expenses, but I'm still not ready to refactor that aspect. The copy and
paste nature of it right now is still (mostly) working for me.